Implement both the Add and Delete operations as nops
These two operations are called when docker tries to store and delete
credentials (docker login and docker logout) respectively. This change
makes it so that both of these operations are implemented as nops
instead of returning a "not implemented" error.

The goal here is to ensure compatibilities with applications and tools
that call docker login or docker logout for the user as part of their
normal operations.